+++ tests/testthat/test-developer-test-toInteger.R
@@ -50,19 +50,13 @@ test_that("Test: toInteger for multiple of 5", {
 test_that("Test: toInteger for survival endpoint event count works properly", {
   # This gives 252.1852 as sample size, 227.1393 as final event count
   x <- gsSurvCalendar(hr = 0.64)
-  # Should round event counts (up for final) and
-  # round up final event count final sample size only as well
+  # Should round event counts down and round up final sample size.
   y <- toInteger(x)
 
-  # Interim event counts rounded
-  expect_equal(min(abs(round(x$n.I[1:x$k - 1]) - y$n.I[1:x$k - 1])), 0)
-  # Final event count round up
-  expect_true(
-    (y$n.I - x$n.I)[x$k] >= 0,
-    info = "Test toInteger of gsSurv rounds up final event count"
-  )
+  # Event counts rounded down
+  expect_equal(y$n.I, floor(x$n.I))
   # Final sample size rounds to even
   expect_true(as.integer((y$eNC + y$eNE)[x$k]) %% 2 == 0)
-  # Final event count rounds up
+  # Final sample size rounds up
   expect_gte((y$eNC + y$eNE - x$eNC - x$eNE)[x$k], 0)
 })